+++ src/__tests__/featureflags.js
@@ -1,9 +1,8 @@
 import { PostHogFeatureFlags } from '../posthog-featureflags'
 
 describe('featureflags', () => {
-    given('properties', () => ({ $override_feature_flags: false, $active_feature_flags: ['beta-feature'] }))
-
     given('instance', () => ({
+        get_config: jest.fn().mockImplementation((key) => given.config[key]),
         get_property: (key) => given.properties[key],
         capture: () => {},
     }))
@@ -14,19 +13,56 @@ describe('featureflags', () => {
         jest.spyOn(given.instance, 'capture').mockReturnValue()
     })
 
+    given('properties', () => ({
+        $active_feature_flags: ['beta-feature', 'alpha-feature-2', 'multivariate-flag'],
+        $enabled_feature_flags: {
+            'beta-feature': true,
+            'alpha-feature-2': true,
+            'multivariate-flag': 'variant-1',
+        },
+        $override_feature_flags: false,
+    }))
+
     it('should return the right feature flag and call capture', () => {
-        expect(given.featureFlags.getFlags()).toEqual(['beta-feature'])
+        expect(given.featureFlags.getFlags()).toEqual(['beta-feature', 'alpha-feature-2', 'multivariate-flag'])
+        expect(given.featureFlags.getFlagVariants()).toEqual({
+            'alpha-feature-2': true,
+            'beta-feature': true,
+            'multivariate-flag': 'variant-1',
+        })
         expect(given.featureFlags.isFeatureEnabled('beta-feature')).toEqual(true)
         expect(given.featureFlags.isFeatureEnabled('random')).toEqual(false)
-        expect(given.instance.capture).toHaveBeenCalledTimes(2)
+        expect(given.featureFlags.isFeatureEnabled('multivariate-flag')).toEqual(true)
+
+        expect(given.instance.capture).toHaveBeenCalledTimes(3)
 
         // It should not call `capture` on subsequent calls
         expect(given.featureFlags.isFeatureEnabled('beta-feature')).toEqual(true)
-        expect(given.instance.capture).toHaveBeenCalledTimes(2)
+        expect(given.instance.capture).toHaveBeenCalledTimes(3)
     })
 
     it('should return the right feature flag and not call capture', () => {
         expect(given.featureFlags.isFeatureEnabled('beta-feature', { send_event: false })).toEqual(true)
         expect(given.instance.capture).not.toHaveBeenCalled()
+    })
+
+    it('supports overrides', () => {
+        given('properties', () => ({
+            $active_feature_flags: ['beta-feature', 'alpha-feature-2', 'multivariate-flag'],
+            $enabled_feature_flags: {
+                'beta-feature': true,
+                'alpha-feature-2': true,
+                'multivariate-flag': 'variant-1',
+            },
+            $override_feature_flags: {
+                'beta-feature': false,
+                'alpha-feature-2': 'as-a-variant',
+            },
+        }))
+        expect(given.featureFlags.getFlags()).toEqual(['alpha-feature-2', 'multivariate-flag'])
+        expect(given.featureFlags.getFlagVariants()).toEqual({
+            'alpha-feature-2': 'as-a-variant',
+            'multivariate-flag': 'variant-1',
+        })
     })
 })